Moved hooks in the data layer. Added Interactors in between.
