+++ src/features/business/domain/interactor/useBusinessDetailsInteractor.ts
@@ -1,0 +1,12 @@
+import { useBusinessDetailsQuery } from '../../data/graphql/hook/useBusinessDetailsQuery';
+import { Business } from '../model/Business';
+
+export interface UseBusinessDetailsInteractorHook {
+  business: Business | undefined;
+  loading: boolean;
+  error: Error | undefined;
+}
+
+export const useBusinessDetailsInteractor = (businessId: String): UseBusinessDetailsInteractorHook => {
+  return useBusinessDetailsQuery(businessId);
+};
